Splitting menu functions to own file
