+++ Makefile.am
@@ -26,9 +26,9 @@ LDADD += -lselinux
 endif
 
 bin_PROGRAMS = $(top_builddir)/bin/show $(top_builddir)/bin/sf
-__top_builddir__bin_show_SOURCES = src/show.c src/showfunctions.c src/showmenus.c src/colors.c src/common.c
+__top_builddir__bin_show_SOURCES = src/show.c src/showfunctions.c src/showmenus.c src/colors.c src/common.c src/menu.c
 __top_builddir__bin_show_LDADD = $(LDADD)
-__top_builddir__bin_sf_SOURCES = src/sf.c src/sfmenus.c src/colors.c src/common.c
+__top_builddir__bin_sf_SOURCES = src/sf.c src/sfmenus.c src/colors.c src/common.c src/menu.c
 __top_builddir__bin_sf_LDADD = $(LDADD)
 dfshowconf_DATA = conf/dfshow.conf
 dfshowdata_DATA = themes/*
