+++ src/show.c
@@ -31,6 +31,7 @@
 #include "showfunctions.h"
 #include "showmenus.h"
 #include "colors.h"
+#include "menu.h"
 #include "common.h"
 #include "show.h"
